+++ gdsfactory/components/mzi_lattice.py
@@ -4,8 +4,7 @@ from typing import Tuple
 from gdsfactory.cell import cell
 from gdsfactory.component import Component
 from gdsfactory.components.coupler import coupler as coupler_function
-from gdsfactory.components.mzi import mzi as mzi_function
-from gdsfactory.components.straight import straight as straight_function
+from gdsfactory.components.mzi import mzi_coupler
 from gdsfactory.types import ComponentFactory
 
 
@@ -14,12 +13,28 @@ def mzi_lattice(
     coupler_lengths: Tuple[float, ...] = (10.0, 20.0),
     coupler_gaps: Tuple[float, ...] = (0.2, 0.3),
     delta_lengths: Tuple[float, ...] = (10.0,),
-    mzi: ComponentFactory = mzi_function,
+    mzi: ComponentFactory = mzi_coupler,
     splitter: ComponentFactory = coupler_function,
-    straight: ComponentFactory = straight_function,
     **kwargs,
 ) -> Component:
     r"""Mzi lattice filter.
+
+    Args:
+        coupler_lengths: list of length for each coupler
+        coupler_gaps: list of coupler gaps
+        delta_lengths: list of length differences
+        mzi: function for the mzi
+        splitter: splitter function
+
+    keyword Args:
+        length_y: vertical length for both and top arms
+        length_x: horizontal length
+        bend: 90 degrees bend library
+        straight: straight function
+        straight_y: straight for length_y and delta_length
+        straight_x_top: top straight for length_x
+        straight_x_bot: bottom straight for length_x
+        cross_section: for routing (sxtop/sxbot to combiner)
 
     .. code::
 
@@ -53,7 +68,6 @@ def mzi_lattice(
         combiner=combiner1,
         with_splitter=True,
         delta_length=delta_lengths[0],
-        straight=straight,
         **kwargs,
     )
     c.add_ports(sprevious.get_ports_list(port_type="electrical"))
@@ -74,7 +88,6 @@ def mzi_lattice(
             combiner=combiner1,
             with_splitter=False,
             delta_length=delta_length,
-            straight=straight,
             **kwargs,
         )
         splitter_settings = combiner_settings
@@ -107,6 +120,7 @@ if __name__ == "__main__":
     # dl0 = [0, 50, 100]
 
     c = mzi_lattice(
-        coupler_lengths=cpl, coupler_gaps=cpg, delta_lengths=dl0, length_x=10
+        coupler_lengths=cpl, coupler_gaps=cpg, delta_lengths=dl0, length_x=1
     )
+    c = mzi_lattice(delta_lengths=(20,), length_x=0)
     c.show()