+++ CHANGELOG.md
@@ -17,6 +17,13 @@ Maybe:
 - mypy passing
 - pass force-regen flag from pytest
 
+## 3.8.1
+
+- add `gf.components.mzi` as a more robust implementation for the MZI
+- rename `gf.components.mzi` to `gf.components.mzi_arms`
+- expose `toolz.compose` as `gf.compose`
+- add `gf.c.mzi1x2`, `mzi1x2_2x2`, `mzi_coupler`
+
 ## 3.8.0
 
 - add `gf.components.copy_layers` to duplicate a component in multiple layers.